Added user for comments.

+++ database/seeders/DatabaseSeeder.php
@@ -19,10 +19,10 @@ class DatabaseSeeder extends Seeder
      */
     public function run()
     {
-        /*if($this->command->confirm('Do you want to refresh the database?')){
+        if($this->command->confirm('Do you want to refresh the database?')){
           $this->command->call('migrate:refresh');
             $this->command->info('Database was refreshed');
-        }*/
+        }
 
         Cache::tags(['blog-post'])->flush();
